+++ src/decisionengine_modules/AWS/tests/test_FigureOfMerit.py
@@ -22,7 +22,7 @@ produces = dict.fromkeys(["AWS_Price_Performance", "AWS_Figure_Of_Merit"], pd.Da
 def create_datablock():
     data_block = {"channel_name": "test"}
     data_block["provisioner_resource_spot_prices"] = pd.read_csv(
-        os.path.join(DATA_DIR, "AWSSpotPriceWithSourceProxy_expected_acquire.csv"), float_precision="high"
+        os.path.join(DATA_DIR, "AWSSpotPriceTransform_expected_acquire.csv"), float_precision="high"
     )
     data_block["Performance_Data"] = (
         pd.read_csv(os.path.join(DATA_DIR, "instance_performance_sample.csv"))
@@ -34,7 +34,7 @@ def create_datablock():
         .drop_duplicates(subset=["AvailabilityZone", "InstanceType"], keep="last")
         .reset_index(drop=True)
     )
-    data_block["AWS_Occupancy"] = pd.read_csv(os.path.join(DATA_DIR, "AWSOcupancyWithSourceProxy_expected_acquire.csv"))
+    data_block["AWS_Occupancy"] = pd.read_csv(os.path.join(DATA_DIR, "AWSOccupancyTransform_expected_acquire.csv"))
     return data_block
